Rename tasks module to workflows
